tests/wpt/web-platform-tests/css/css-scroll-snap/snap-area-capturing-remove-scroll-container.html

<!DOCTYPE html>
<title>
  When an element no longer captures snap positions (e.g., no longer
  scrollable), then its currently captured snap areas must be reassigned.
</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap/#captures-snap-positions"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
div {
  position: absolute;
  margin: 0px;
}

html {
  scroll-snap-type: y mandatory;
}

body {
  margin: 0px;
}

#middle-scroller {
  top: 100px;
  height: 500px;
  width: 500px;
  overflow: scroll;
  background-color: rgb(12, 61, 2);
  scroll-snap-type: none;
}

#inner-scroller {
  top: 200px;
  height: 400px;
  width: 400px;
  overflow: scroll;
  background-color: rgb(65, 139, 50);
  scroll-snap-type: y mandatory;
}

.space {
  width: 2000px;
  height: 2000px;
}

#inner-snap-area {
  top: 300px;
  width: 200px;
  height: 200px;
  background-color: blue;
  scroll-snap-align: start;
}

#document-snap-area {
  top: 500px;
  width: 200px;
  height: 200px;
  background-color: lightblue;
  scroll-snap-align: start;
}

</style>
<div class="space"></div>
  <div id="middle-scroller">
    <div class="space"></div>
    <div id="inner-scroller">
        <div class="space"></div>
        <div id="inner-snap-area"></div>
    </div>
  </div>
</div>
<div id="document-snap-area"></div>
<script>

// This tests that making a snap container no longer scrollable will reassign
// its snap areas to the next scrollable ancestor, per spec [1].
// [1] https://drafts.csswg.org/css-scroll-snap/#captures-snap-positions
test(() => {
  const innerscroller = document.getElementById("inner-scroller");
  const middlescroller = document.getElementById("middle-scroller");
  const documentscroller = document.scrollingElement;

  // Middle scroller doesn't snap.
  // Document scroller should snap to its only captured area.
  documentscroller.scrollBy(0, 100);
  middlescroller.scrollBy(0, 10);
  assert_equals(innerscroller.scrollTop, 0);
  assert_equals(middlescroller.scrollTop, 10);
  assert_equals(documentscroller.scrollTop, 500);

  // Inner scroller snaps.
  innerscroller.scrollBy(0, 10);
  assert_equals(innerscroller.scrollTop, 300);
  assert_equals(middlescroller.scrollTop, 10);
  assert_equals(documentscroller.scrollTop, 500);

  // Inner scroller is no longer a scroll container.
  innerscroller.style.setProperty("overflow", "visible");
  assert_equals(innerscroller.scrollTop, 0);
  assert_equals(middlescroller.scrollTop, 10);
  assert_equals(documentscroller.scrollTop, 500);

  // The new snap container is the middle scroller, which has snap-type 'none'.
  // Per spec, the scroll container should capture snap positions even if it has
  // snap-type 'none'.
  // The middle scroller should not snap.
  // The document scroller should still only snap to its captured snap area.
  documentscroller.scrollBy(0, 100);
  middlescroller.scrollBy(0, 10);
  assert_equals(innerscroller.scrollTop, 0);
  assert_equals(middlescroller.scrollTop, 20);
  assert_equals(documentscroller.scrollTop, 500);

  // The scroll container should now be at the document level.
  middlescroller.style.setProperty("overflow", "visible");
  documentscroller.scrollBy(0, -10);
  assert_equals(innerscroller.scrollTop, 0);
  assert_equals(middlescroller.scrollTop, 0);

  // Check that the existing snap area did not get removed when reassigning
  // the inner snap area.
  assert_equals(documentscroller.scrollTop, 500);

  // Check that the inner snap area got reassigned to the document.
  documentscroller.scrollBy(0, 150);
  assert_equals(documentscroller.scrollTop, 600);
}, 'Making a snap container not scrollable should promote the next scrollable\
 ancestor to become a snap container.');
</script>
